fix(docbuild): remaining of 134 batch
Of the first 134 Doxygen warnings that are not "is not documented"
warnings, a batch of 28 or so files were already updated to have
corrected `@file` commands.  The remaining 106 warnings are handled
in this commit.

Next step will be addressing the "is not documented" warnings that are
legitimate Doxygen docstring errors.

Fixes part 2 of issue #6480.